make it smoooother

+++ js/debug.js
@@ -91,7 +91,7 @@ function debugSetVector() {
 
 function debugMaxVelocity() {
   var vals = Utils.extractNameValues( document.getElementById('debug_phy'));
-  vals = parseInt(vals["setVelocity"]);
+  vals = parseFloat(vals["setVelocity"]);
   for (var Obj of INIT_PLAY_AREA.objects) {
     Obj.max_vector = vals;
   }
